+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.10.4-wip.13
 - :arrow_up: Added specs for Messages::Video::Transcript service
+- :arrow_up: Removed transcription field from messages#create
 
 ### 2.10.4-wip.12
 - :arrow_up: Fixed video transcription persistence